pyproject.toml
[tool.poetry]
name = "code-test"
version = "0.1.0"
description = "app to satisfy the LDCI code-test"
authors = ["maulanayos <imamaulanayos@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.1"
pillow = "^11.2.1"
djangorestframework = "^3.16.0"
django-rest-knox = "^5.0.2"
djangorestframework-simplejwt = "^5.5.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
